+++ contracts/ModuleRegistry.sol
@@ -4,6 +4,7 @@ import "./interfaces/IModuleRegistry.sol";
 import "./interfaces/IModuleFactory.sol";
 import "./interfaces/ISecurityToken.sol";
 import "./interfaces/ISecurityTokenRegistry.sol";
+import "./interfaces/IRegistry.sol";
 import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
 
 /**
@@ -12,7 +13,7 @@ import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
 * Anyone can register modules, but only those "approved" by Polymath will be allowed to everyone.
 */
 
-contract ModuleRegistry is IModuleRegistry, Ownable {
+contract ModuleRegistry is IModuleRegistry, Ownable, IRegistry {
 
     // Mapping used to hold the type of module factory corresponds to the address of the Module factory contract
     mapping (address => uint8) public registry;