+++ migrations/2_deploy_contracts.js
@@ -9,6 +9,7 @@ const TickerRegistry = artifacts.require('./TickerRegistry.sol')
 const STVersionProxy001 = artifacts.require('./tokens/STVersionProxy001.sol')
 const DevPolyToken = artifacts.require('./helpers/PolyToken.sol')
 const cappedSTOSetupCost = 20000 * Math.pow(10,18);   // 20K POLY fee
+const initRegFee = 250 * Math.pow(10, 18);       // 250 POLY fee for registering ticker or security token in registry
 let PolyToken
 
 const Web3 = require('web3')
@@ -98,10 +99,10 @@ module.exports = function (deployer, network, accounts) {
     return deployer.deploy(STVersionProxy001, GeneralTransferManagerFactory.address, {from: PolymathAccount})
     }).then(() => {
       // I) Deploy the TickerRegistry Contract (It is used to store the information about the ticker)
-    return deployer.deploy(TickerRegistry, {from: PolymathAccount})
+    return deployer.deploy(TickerRegistry, PolyToken, initRegFee, {from: PolymathAccount})
     }).then(() => {
       // J) Deploy the SecurityTokenRegistry contract (Used to hold the deployed secuirtyToken details. It also act as the interface to deploy the SecurityToken)
-    return deployer.deploy(SecurityTokenRegistry, PolyToken, ModuleRegistry.address, TickerRegistry.address, STVersionProxy001.address, {from: PolymathAccount})
+    return deployer.deploy(SecurityTokenRegistry, PolyToken, ModuleRegistry.address, TickerRegistry.address, STVersionProxy001.address, initRegFee, {from: PolymathAccount})
     }).then(() => {
     return TickerRegistry.deployed().then((tickerRegistry) => {
       // K) SecurityTokenRegistry address make available to the TickerRegistry contract for accessing the securityTokenRegistry functions